update check mark in reviewviewcontroller and remove same code in reviewimageviewcontroller
